requirements.txt
cachetools
certifi
charset-normalizer
google-api-core
google-auth
google-auth-httplib2
google-auth-oauthlib
httplib2
idna
oauthlib
pip
proto-plus
protobuf
pyasn1
pyasn1_modules
pyparsing
requests
requests-oauthlib
rsa
uritemplate
urllib3
python-dotenv
